krunner: Use layer shell on wayland
Removes the need for asking Plasma for struts via dbus, which slows krunner
significantly down since we wait for the response before showing. With layer
shell kwin can place the window correctly. On X this  works correctly via struts
(as it did before the introduction of the dbus call).
Care is needed with regards to PlasmaQuick::Dialog which View inherits from,
some events are not forwarded to it to not create a plasmashell surface.
A setGeometry call is needed so the window gets correct borders and shadows,
in the future Dialog shouls use anchors to determine those.